create 'complete' package and provide a way to customize the autocompletion mode
